+++ acp/identity/identity.go
@@ -161,9 +161,5 @@ func didFromPublicKey(publicKey *secp256k1.PublicKey, producer didProducer) (str
 
 // IntoRawIdentity converts an `Identity` into a `RawIdentity`.
 func (identity Identity) IntoRawIdentity() RawIdentity {
-	return RawIdentity{
-		PrivateKey: hex.EncodeToString(identity.PrivateKey.Serialize()),
-		PublicKey:  hex.EncodeToString(identity.PublicKey.SerializeCompressed()),
-		DID:        identity.DID,
-	}
+	return newRawIdentity(identity.PrivateKey, identity.PublicKey, identity.DID)
 }
